Change wait_for_boot method for node availability
The current wait_for_boot method uses a socket connection to port 22
to determine whether commands can be run via SSH on a given node.
However this is not definitive as even if a connection to port 22
can be made, SSH may not be immediately available. A more robust
approach is to repeatedly attempt a basic command until it executes
successfully.

Change-Id: Ib80e29f947a47e1b5349236a7e9db42b59c09b10